v3.1.5. Reversed posts array in profile page. Updated styles in profile gallery
