+++ frontend/src/pages/Profile/ProfileEntities.tsx
@@ -2,7 +2,8 @@ import React from "react";
 import { Grid, Typography } from "@mui/material";
 import { ProfileDto } from "../../types/profile/profile.dto";
 import { Link } from "react-router-dom";
-
+// @ts-ignore
+import styles from '../Login/Login.module.css';
 interface ProfileEntitiesInterface {
 	profile: ProfileDto;
 }
@@ -11,7 +12,7 @@ export const ProfileEntities: React.FC<ProfileEntitiesInterface> = ({ profile })
 
 	return (
 		<Grid>
-			<Typography variant="h6"><Link to={'/profile/'+profile.id+'?gallery=true'}>Photos</Link></Typography>
+			<Typography variant="h6"><Link to={'/profile/'+profile.id+'?gallery=true'} className={styles.link}>Photos</Link></Typography>
 		</Grid>
 	);
 };
